+++ platzky/admin/templates/login.html
@@ -13,7 +13,7 @@
   <div class="row align-items-center">
 
 {% for login_method in login_methods %}
-    {{ login_method }}
+    {{ login_method | safe }}
 {% endfor %}
   </div>
 </div>
