+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "platzky"
-version = "0.3.0"
+version = "0.3.1"
 description = "Not only blog engine"
 authors = []
 license = "MIT"
